+++ packages/plugins/eslint-plugin-react-naming-convention/src/configs/recommended.ts
@@ -6,5 +6,4 @@ export const rules = {
   "react-naming-convention/context-name": "warn",
   "react-naming-convention/id-name": "warn",
   "react-naming-convention/ref-name": "warn",
-  "react-naming-convention/use-state": "warn",
 } as const satisfies Record<string, RuleConfig>;