+++ packages/plugins/eslint-plugin/src/configs/x.ts
@@ -29,6 +29,8 @@ export const rules = {
   "@eslint-react/no-direct-mutation-state": "error",
   "@eslint-react/no-forward-ref": "warn",
   "@eslint-react/no-missing-key": "error",
+  "@eslint-react/use-memo": "error",
+  "@eslint-react/use-state": "warn",
   "@eslint-react/no-nested-component-definitions": "error",
   "@eslint-react/no-nested-lazy-component-declarations": "error",
   "@eslint-react/no-redundant-should-component-update": "error",